Some changes to the TimeCapsule screen: 1) You can now see the user who left the capsule 2) clicking on the user's name will take you to their profile 3) Number of views and number of distinct viewers is now functional 4) Modifications to the way timestamps show up for the capsule and comments
Other changes:
I think everything is using JSON now, let me know if something is not

Known problems:
I am getting a ConcurrentModificationException when the map is being drawn on
This only happens sporadically, I believe it is because drawUser() and mapView.invalidate() are trying to draw on the mapOverlays at the same time. I have not fixed the problem yet, but hopefully will soon.